Add image tracking support
Track container images in addition to hosts/servers.

Container images are typically static, but the data model also allows to
track upgradable packages; this could be implemented by spinning up a
instances with all container images in use on a daily basis and
ingesting package updates.

Updates for images need to be submitted from a system whitelisted
for data submission in PROXY_HOSTS.

This also adds a separate image data submission mode to the CLI
frontend.

Bug: T237978
Change-Id: Ia0b86ef3af715da89afab455f64b90a172f0806a