Create debmonitor user on buster with adduser
A glibc bug in Buster crashes systemd-sysusers if /etc/gshadow contains an
entry with more than 1024 chars. This is the case on e.g. the bastion hosts
which populate a group with all users.

As such, continue to use adduser on Buster for now (the fix in glibc might
still get backported to a point release)

Bug: T256098
Change-Id: Id74d4423f2e195e075f7cfa662ad84c067c98b40