Changes to Python infrastucture to help building Debian package.
The Debian packaging uses the Python build/setup.py infrastructure
to build the Debian package and find dependencies. This adds the
changes required to build the Debmonitor server as a .deb, rather
than building the client.

Changes to the Debian packaging configuration follows in a separate
patch to the "debian" branch.

Change-Id: Ia4537847d66eacb17b6c8227388c264be0ce2332