Add MANIFEST.in
To get template, css, JavaScript and image included in the Debian
source package, which will in turn get them included in the binary
package, we need a MANIFEST.in file.

Without this file, setuptools won't include the data files in the sdist,
which means it also can't find it later when trying to build the binary
package.

Change-Id: Iea96424d4c4a99691fd5ddb4477f38d80ebcf65f