Allow to save new OS names without them being present on the DB
We recently changed the debmonitor-cli code to send OS names
with version-id ('Debian 11' or 'Ubuntu 22.04' for example),
but we had to insert the new allowed values manually on the DB
to avoid errors.
This change adds more logic to the views.py code, in order to allow
new OS name values if they match a certain regex. This should allow
us to handle new OS versions without necessarily remember to update
the DB manually.

The new logic is split into two part:
* A validation function added to the OS model, that we can call
  via "clean_fields()". The validation regex is stored in a single
  place, even if the code to test it is replicated into both views.py
  (I didn't find a better way to do it).
* New code in views.py to be able to handle a OS.DoesNotExist
  exception. It does validate the new OS name and return an exception only
  if it doesn't follow our regex, or it just add the new value to the DB
  and proceeds.

I am not a Django expert so I am not 100% sure that this is the right
way to proceed, but all the test suite works fine.

Related debmonitor-cli change: https://gerrit.wikimedia.org/r/1043780

Bug: T368744
Change-Id: I45c6da2d5db33bf664fb744f2e36958802419775