Unvendor Bootstrap
Now installed from Debian. I've also dropped the integrity links, they are
useful if one fetches JS/CSS from a CDN, but here in our case the files
from a trusted source. And otherwise we'd need to update the hashes after
there are security updates in Debian.

The new JAVASCRIPT_STORAGE option defaults to 'Debian' to use bootstrap
from Debian and can be set to an arbitrary different value to fetch
it from a CDN.

Bug: T397696
Change-Id: I501147e5b8549500efe871d4c15ca6c73f2ff2a8